added some decimals

+++ VSharp.SILI.Core/Types.fs
@@ -271,7 +271,7 @@ module internal Types =
     let UInt64 = Numeric typeof<uint64>
     let F = Numeric typeof<float>
     let D = Numeric typeof<double>
-    let Dec = Numeric typeof<double>
+    let Dec = Numeric typeof<Decimal>
     
     let (|StringType|_|) = function
         | typ when typ = String -> Some()
